move codegen into shared area
