--- packages/api/codegen.yml
@@ -1,17 +1,0 @@
----
-schema:
-  - http://localhost:3000/v1/graphql
-documents:
-  - './src/**/*.gql'
-overwrite: true
-generates:
-  './src/generated/graphql.ts':
-    plugins:
-      - typescript
-      - typescript-operations
-      - typescript-graphql-request
-    config:
-      skipTypename: true
-  './graphql.schema.json':
-    plugins:
-      - introspection
